// src/api/models/VendorRegisterModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VendorRegisterModel = {
  firstName?: string | null;
  lastName?: string | null;
  email?: string | null;
  password?: string | null;
  confirmPassword?: string | null;
  phoneNumber?: string | null;
  isStudent?: boolean;
  university?: string | null;
  businessName?: string | null;
  businessDescription?: string | null;
  businessAddress?: string | null;
};